requirements.txt
django
celery
redis
djangorestframework

